removed if statement

+++ modules/assim.sequential/R/sda.enkf_MultiSite.R
@@ -212,7 +212,7 @@ sda.enkf.multisite <- function(settings, obs.mean, obs.cov, Q = NULL, restart = 
     if (t > 1){
       #removing old simulations
       ## keep all old simulations right now since there is debugging going on
-      if (!(keepNC))
+      if (!(keepNC) && t != 1)
       {
         unlink(list.files(outdir, "*.nc", recursive = TRUE, full.names = TRUE))
       }
@@ -511,9 +511,7 @@ sda.enkf.multisite <- function(settings, obs.mean, obs.cov, Q = NULL, restart = 
     #writing down the image - either you asked for it or not :)
 
     if (t%%2==0 || t==nt && control$TimeseriesPlot)  post.analysis.multisite.ggplot(settings, t, obs.times, obs.mean, obs.cov, FORECAST, ANALYSIS ,plot.title=control$plot.title, facetg=control$facet.plots, readsFF=readsFF)
-    if (t == 1){
-      unlink(list.files(outdir, "*.nc", recursive = TRUE, full.names = TRUE))
-    }
+   
   } ### end loop over time
   
 } # sda.enkf
